Create stake-hbar.md (#304)
* Create stake-hbar.md

Creating this page to better showcase quickly how retail can participate in staking, without reading through the full gauntlet of information about the phases, etc. 

Could we create a folder for "Staking" to put this into, along with the details of the program in the other one?

* Update stake-hbar.md

Missed adding information about triggering rewards.